+++ providers/eks/standard/eks_suite_test.go
@@ -6,10 +6,9 @@ import (
 
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 	. "github.com/onsi/gomega"    //nolint:staticcheck
-	cr "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/cluster-standup-teardown/pkg/clusterbuilder/providers/capa"
-	"github.com/giantswarm/clustertest/pkg/client"
+	clustertestclient "github.com/giantswarm/clustertest/pkg/client"
 	"github.com/giantswarm/clustertest/pkg/wait"
 
 	"github.com/giantswarm/cluster-test-suites/internal/state"
@@ -17,9 +16,9 @@ import (
 )
 
 func TestEKSStandard(t *testing.T) {
-	suite.Setup(false, &capa.ManagedClusterBuilder{}, func(client *client.Client) {
+	suite.Setup(false, &capa.ManagedClusterBuilder{}, func(client *clustertestclient.Client) {
 		Eventually(
-			wait.AreNumNodesReady(state.GetContext(), client, 2, &cr.MatchingLabels{"node-role.kubernetes.io/worker": ""}),
+			wait.AreNumNodesReady(state.GetContext(), client, 2, clustertestclient.DoesNotHaveLabels{"node-role.kubernetes.io/control-plane"}),
 			20*time.Minute, 15*time.Second,
 		).Should(BeTrue())
 	})